Still migrating compute() interfaces

+++ examples/test_memory.cc
@@ -138,7 +138,7 @@ size_t readNumber(std::istream &in)
   if (('0' > c) || ('9' < c)) {
     parseError(0, "number expected");
   }
-  num = c - '0';
+  num = size_t(c - '0');
   for (;;) {
     c = in.get();
     if (!in) break;
@@ -147,7 +147,7 @@ size_t readNumber(std::istream &in)
       break;
     }
     num *= 10;
-    num += c - '0';
+    num += size_t(c - '0');
   }
 #ifdef DEBUG_PARSER
   std::cerr << "Consumed number " << num << "\n";
